MINIFICPP-1676 - Refactor, add tests
