+++ libminifi/src/core/extension/ExtensionManager.cpp
@@ -20,12 +20,11 @@
 #include <algorithm>
 
 #include "core/logging/LoggerConfiguration.h"
-#include "utils/file/FileUtils.h"
 #include "core/extension/Executable.h"
 #include "utils/file/FilePattern.h"
 #include "core/extension/DynamicLibrary.h"
-#include "utils/file/FileView.h"
 #include "agent/agent_version.h"
+#include "core/extension/Utils.h"
 
 namespace org {
 namespace apache {
@@ -33,90 +32,6 @@ namespace nifi {
 namespace minifi {
 namespace core {
 namespace extension {
-
-namespace {
-
-struct Timer {
-  explicit Timer(std::function<void(int)> cb): cb_(std::move(cb)) {}
-  ~Timer() {
-    auto end = std::chrono::steady_clock::now();
-    int elapsed = gsl::narrow<int>(std::chrono::duration_cast<std::chrono::milliseconds>(end - start_).count());
-    cb_(elapsed);
-  }
-  std::chrono::steady_clock::time_point start_{std::chrono::steady_clock::now()};
-  std::function<void(int)> cb_;
-};
-
-struct LibraryDescriptor {
-  std::string name;
-  std::filesystem::path dir;
-  std::string filename;
-
-  [[nodiscard]]
-  bool verify(const std::shared_ptr<logging::Logger>& logger) const {
-    auto path = getFullPath();
-    Timer timer{[&] (int ms) {
-      logger->log_error("Verification for '%s' took %d ms", path.string(), ms);
-    }};
-    try {
-      utils::file::FileView file(path);
-      const std::string_view begin_marker = "__EXTENSION_BUILD_IDENTIFIER_BEGIN__";
-      const std::string_view end_marker = "__EXTENSION_BUILD_IDENTIFIER_END__";
-      auto build_id_begin = std::search(file.begin(), file.end(), begin_marker.begin(), begin_marker.end());
-      if (build_id_begin == file.end()) {
-        logger->log_error("Couldn't find start of build identifier in '%s'", path.string());
-        return false;
-      }
-      std::advance(build_id_begin, begin_marker.length());
-      auto build_id_end = std::search(build_id_begin, file.end(), end_marker.begin(), end_marker.end());
-      if (build_id_end == file.end()) {
-        logger->log_error("Couldn't find end of build identifier in '%s'", path.string());
-        return false;
-      }
-      std::string build_id(build_id_begin, build_id_end);
-      if (build_id != AgentBuild::BUILD_IDENTIFIER) {
-        logger->log_error("Build identifier does not match in '%s', expected '%s', got '%s'", path.string(), AgentBuild::BUILD_IDENTIFIER, build_id);
-        return false;
-      }
-    } catch (const std::ios_base::failure& file_error) {
-      logger->log_error("Error while verifying library '%s': %s", path.string(), file_error.what());
-      return false;
-    }
-    return true;
-  }
-
-  [[nodiscard]]
-  std::filesystem::path getFullPath() const {
-    return dir / filename;
-  }
-};
-
-std::optional<LibraryDescriptor> asDynamicLibrary(const std::filesystem::path& path) {
-#if defined(WIN32)
-  const std::string extension = ".dll";
-#elif defined(__APPLE__)
-  const std::string extension = ".dylib";
-#else
-  const std::string extension = ".so";
-#endif
-
-#ifdef WIN32
-  const std::string prefix = "";
-#else
-  const std::string prefix = "lib";
-#endif
-  std::string filename = path.filename().string();
-  if (!utils::StringUtils::startsWith(filename, prefix) || !utils::StringUtils::endsWith(filename, extension)) {
-    return {};
-  }
-  return LibraryDescriptor{
-    filename.substr(prefix.length(), filename.length() - extension.length() - prefix.length()),
-    path.parent_path(),
-    filename
-  };
-}
-
-}  // namespace
 
 const std::shared_ptr<logging::Logger> ExtensionManager::logger_ = logging::LoggerFactory<ExtensionManager>::getLogger();
 
@@ -152,7 +67,7 @@ bool ExtensionManager::initialize(const std::shared_ptr<Configure>& config) {
       logger_->log_error("Error in subpattern '%s': %s", std::string{subpattern}, std::string{error_msg});
     }));
     for (const auto& candidate : candidates) {
-      auto library = asDynamicLibrary(candidate);
+      auto library = internal::asDynamicLibrary(candidate);
       if (!library || !library->verify(logger_)) {
         continue;
       }
